Extract more tgsl helpers
